Bug fixing, period sending increased and energy consumption percentages shown.
1. Fixed bug about data representation of Co2 sensing.

2. All power energy info are printed out. Change sending period from 5
seconds to 60 seconds. With this new configuration the more energy
consumed component is the radio. In order to reduce this wasting, is
necessary to reduce (dramatically) the duty cycle (if possible).
Moreover is needed to change the current ContikiMac protocol with the
802.15.4 protocol because the real device supports this standard.

2b. For above reason the percentage of energy consumption of each module
is shown.

+++ sendPacket2/energy/energy_data_structure.h
@@ -2,7 +2,7 @@
 #define ENERGY_CONS_H 3424
 struct energy_cons{
 
-unsigned long time_sensing;
+uint16_t time_sensing;
 //cpu energy cons
 unsigned long cpu;
 unsigned long lpm;
